misc: chore: Collapse XCIFileTrimmerLog implementations into a single class.
